Add core package and move the 2 tests to this package
